pyproject.toml
[project]
name = "repobot"
version = "0.0.1"
description = "A bot to help maintain an open source GitHub repository."
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["repobot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
